requirements.txt
torch
numpy
tqdm
plotly
pandas
einops
ipython
sae-lens
transformer-lens
python-dotenv
transformers
datasets
pynvml
torcheval
